+++ src/pages/User/Mylist/Details.tsx
