+++ src/pages/User/Mylist/Registrations.tsx
